requirements.txt
Django>=5.2
psycopg2>=2.9
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-dotenv>=1.0
# Optional future extensions
# djongo>=1.3.6
# mongoengine>=0.27
# neo4j>=5.13
# py2neo>=2021.2.3

